+++ content/notes/books/novel/article-1.md
