+++ content/notes/books/novel/article-2.md
