+++ content/notes/books/novel/article-3.md
