+++ content/notes/books/novel/article-4.md
