+++ content/notes/word/motto/article-2.md
