+++ content/notes/word/wise/article-2.md
