--- content/others/word/wise/article-2.md
@@ -1,10 +1,0 @@
----
-title: Introduction
-description: Learn how to use @nuxt/content.
-tags: 
-  - test1
-  - test2
----
-
-## 見出し２
-### 見出し3
