--- pages/others/_catergory/_family/_slug.vue
@@ -1,33 +1,0 @@
-<template>
-  <DefaultTemplate :isPostsPage="true">
-    <nuxt-link :to="'/others/' + catergory + '/' + family">⇦{{catergory}}/{{family}}の記事一覧へ戻る</nuxt-link>
-    <p>ファミリースラッグ</p>
-    <h2>{{ article.title }}</h2>
-    <nuxt-content :document="article" />
-  </DefaultTemplate>
-</template>
-
-<script>
-import DefaultTemplate from "@/components/Templates/defaulttemplate.vue";
-export default {
-  components: {
-    DefaultTemplate,
-  },
-  async asyncData({ $content, params, route, error }) {
-    const { catergory, family, slug } = params;
-    let article;
-
-    try {
-      article = await $content('others', catergory, family, slug).fetch();
-    } catch (e) {
-      error({ message: "others-family-data not found" });
-    }
-
-    return {
-      article,
-      catergory,
-      family
-    };
-  },
-};
-</script>
